Add chat is now fully networked
The pysocket.py module now is fully functioning with client-server connection  as well as a nickname system and a main menu. The overall state of the project however is messy since this was done as a challenge of how fast I could get it functioning on a base level to myself.